Create movement_time_n
Generates x and y positions for all time points of all robots. The positions for each robot are stored in vectors and the vectors are stored in a cell. The x and y movements can be generated with different random variate generations, but the range of movement in any directions should stay within -50 and 50 units.