Create SSATestTauLeap
This is the main program of a practice SSA. It includes 3 chemical reactions and 3 substances. The number of simulations and the maximum time of each simulation is controlled by the user. Tau leaping is used to generate time steps. The amounts of each substance are plotted on three separate graphs.