Create InitializeParameters
This function defines the number of reactions and the number of species. It controls the reaction rate of each reaction and the amount of each species consumed in each reaction. It sets the initial time for the reactions (0 seconds). Additionally, it sets the initial amount of each species. 

The initial amount of each species, the reaction rates and the amount of each species consumed in each reaction can be changed. If the number of species or reactions changes, additional changes would need to be made in the main program.